Service tests

+++ Backend-Spring/src/main/java/pl/danielpiskorz/thesmartway/service/TaskService.java
@@ -3,6 +3,6 @@ package pl.danielpiskorz.thesmartway.service;
 import pl.danielpiskorz.thesmartway.model.Task;
 
 public interface TaskService {
-	Task incrementTodoIndex(String username, int taskId);
+	Task incrementTodoIndex(String username, long taskId);
 	Task updateTodos(String username, Task task);
 }
